+++ Sources/ZcashLightClientKit/Synchronizer/SDKSynchronizer.swift
@@ -665,6 +665,16 @@ public class SDKSynchronizer: Synchronizer {
         try await blockProcessor.getCustomUnifiedAddress(accountUUID: accountUUID, receivers: receivers)
     }
 
+    // MARK: Rescan
+
+    public func rescanFrom(height: BlockHeight) async throws {
+        let checkpointSource = initializer.container.resolve(CheckpointSource.self)
+
+        let checkpoint = checkpointSource.birthday(for: height)
+
+        try await initializer.rustBackend.rewindToChainState(chainState: checkpoint.treeState())
+    }
+    
     // MARK: Rewind
 
     public func rewind(_ policy: RewindPolicy) -> AnyPublisher<Void, Error> {